stocks_products/requirements.txt
aiosmtplib==1.1.6
app==0.0.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
aVKapi==0.1.2
beautifulsoup4==4.11.1
bs4==0.0.1
caugetch==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
clipboard==0.0.4
colorama==0.4.4
commonmark==0.9.1
Django==4.0.6
djangorestframework==3.13.1
getpass4==0.0.14.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
nose==1.3.7
packaging==21.3
parameterized==0.8.1
Pillow==9.2.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.9.1
Pygments==2.12.0
pyngrok==5.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyTelegramBotAPI==4.6.0
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.27.1
rich==10.16.2
secure-smtplib==0.1.1
smtp-mail==0.0.2
smtp-send==0.0.7
socket-api2==0.2.0
socket.py==0.1.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlalchemy-dbml==0.1.0
sqlalchemy-function==0.0.9
sqlalchemy-mutable==1.0.2
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.4
sqlparse==0.4.2
tdqm==0.0.1
threading-tools==0.1.0
tomli==2.0.1
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
vk==3.0
vk-api==11.9.9
VkApiPy==1.0.0.3